change some endpoint name (#23)
